web/pyproject.toml
[project]
name = "web"
version = "0.1.0"
description = "Web application to show information about Vancouver from latest social media and user-submitted data"
authors = [
    {name = "Andrés Ignacio Torres",email = "dev@aitorres.com"}
]
requires-python = ">=3.11"
dependencies = [
    "pika (>=1.3.2,<2.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "sentry-sdk[flask] (>=2.33.0,<3.0.0)",
    "pytest-mock (>=3.14.1,<4.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pre-commit = "^4.2.0"
flake8 = "^7.3.0"
black = "^25.1.0"
mypy = "^1.16.1"
pylint = "^3.3.7"
isort = "^6.0.1"
types-psycopg2 = "^2.9.21.20250718"

[tool.isort]
src_paths = ["web", "tests"]
profile = "black"
line_length = 90

[tool.mypy]
warn_return_any = true
warn_unused_configs = true

[tool.pylint.FORMAT]
max-line-length=90

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
